moved level logics to single location
