+++ packages/warriorjs-cli/src/level/components/constructSeperator.js
