+++ packages/warriorjs-cli/src/level/components/failure.js
