+++ packages/warriorjs-cli/src/level/components/floorMap.js
