+++ packages/warriorjs-cli/src/level/components/getUnitStyle.js
