+++ packages/warriorjs-cli/src/level/components/logMessage.js
