+++ packages/warriorjs-cli/src/level/components/success.js
