+++ packages/warriorjs-cli/src/level/components/totalScore.js
