+++ packages/warriorjs-cli/src/level/components/warriorStatus.js
