+++ packages/warriorjs-cli/src/level/layout/base.js
